slave plugin: Only try once to change nick to master
New girc behaviour is that nick = foo will try foo|1, etc if foo is taken.
Normally (including in other places in this plugin) this is what we want.
But in this specific case, we want to try, but do nothing if we fail.
Otherwise we're changing from one incremented nick to another, which is just spam.